+++ src/lib/__tests__/_validation_data.ts
@@ -239,6 +239,14 @@ export const supportedRecord: IsoRecord = deepMergeObjects(
           role: ['author'],
         },
       ],
+      graphic_overviews: [
+        {
+          identifier: 'overview',
+          description: 'General overview of resource',
+          href: 'https://example.com/image.png',
+          mime_type: 'image/png',
+        },
+      ],
       other_citation_details:
         'Required citation:\n> Gerrish, L. (2024). _x_ (Version 1) [Data set]. NERC EDS UK Polar Data Centre. https://doi.org/10.5285/25585848-7b80-42f7-8d4a-069d7479c287',
     },